refactor: implement new code structure
