+++ api/admins/organization.py
@@ -1,0 +1,75 @@
+
+from django.contrib import admin
+
+from .base import BaseModelAdmin, BaseModelResource, RESOURCE_FIELDS
+from api.models import Department, Tribe, Chapter, Team, Committee
+
+@admin.register(Department)
+class DepartmentAdmin(BaseModelAdmin):
+    list_display = ("name", "date_created", "date_updated")
+    fields = ("uuid", "name", "description", ("date_created", "date_updated"))
+    readonly_fields = ("uuid", "date_created", "date_updated")
+    ordering = ("-date_created", "name")
+    search_fields = ["name"]
+    search_help_text = "جستجو در نام دپارتمان"
+
+
+@admin.register(Tribe)
+class TribeAdmin(BaseModelAdmin):
+    class TribeResource(BaseModelResource):
+        leader = RESOURCE_FIELDS["leader"]
+        department = RESOURCE_FIELDS["department"]
+
+        class Meta:
+            model = Tribe
+            fields = ("name", "leader", "department")
+
+    resource_class = TribeResource
+    list_display = ("name", "department", "leader", "date_created", "date_updated",)
+    fields = ("uuid","name", ("department", "leader"), "description", ("date_created", "date_updated"),)
+    readonly_fields = ("uuid", "date_created", "date_updated")
+    ordering = ("-date_created", "name")
+    search_fields = ["name", "department__name", "leader__name", "leader__email"]
+    search_help_text = "جستجو در نام قبیله، نام دپارتمان، نام لیدر، ایمیل لیدر "
+
+
+@admin.register(Chapter)
+class ChapterAdmin(BaseModelAdmin):
+    list_display = ("name", "department", "leader", "date_created", "date_updated",)
+    fields = ("uuid", "name", ("department", "leader"), "description", ("date_created", "date_updated"),)
+    readonly_fields = ("uuid", "date_created", "date_updated")
+    ordering = ("-date_created", "name")
+    search_fields = ["name", "department__name", "leader__name", "leader__email"]
+    search_help_text = "جستجو در نام چپتر، نام دپارتمان، نام لیدر، ایمیل لیدر "
+
+
+@admin.register(Team)
+class TeamAdmin(BaseModelAdmin):
+    class TeamResource(BaseModelResource):
+        leader = RESOURCE_FIELDS["leader"]
+        department = RESOURCE_FIELDS["department"]
+        tribe = RESOURCE_FIELDS["tribe"]
+
+        class Meta:
+            model = Team
+            fields = ("name", "leader", "department", "tribe")
+
+    resource_class = TeamResource
+    list_display = ("name", "department", "leader", "tribe", "date_created", "date_updated",)
+    fields = ("uuid", "name", ("department", "leader", "tribe"), "description", ("date_created", "date_updated"),)
+    readonly_fields = ("uuid", "date_created", "date_updated")
+    ordering = ("-date_created", "name")
+    search_fields = ["name", "department__name", "leader__name", "leader__email"]
+    search_help_text = "جستجو در نام تیم، نام دپارتمان، نام لیدر، ایمیل لیدر "
+
+
+@admin.register(Committee)
+class CommitteeAdmin(BaseModelAdmin):
+    list_display = ("name", "date_created", "date_updated",)
+    fields = ("uuid", "name", "description", "members", ("date_created", "date_updated"),)
+    filter_horizontal = ("members",)
+    readonly_fields = ("uuid", "date_created", "date_updated")
+    ordering = ("-date_created", "name")
+    search_fields = ["name"]
+    search_help_text = "جستجو در نام کمیته"
+
