+++ api/models/base.py
@@ -2,6 +2,8 @@ import uuid
 
 from django.db import models
 
+__all__ = ['MerlinBaseModel']
+
 class MerlinBaseModel(models.Model):
     uuid = models.UUIDField(unique=True, default=uuid.uuid4, editable=False)
     date_created = models.DateTimeField(
