+++ api/serializers/__init__.py
@@ -1,0 +1,16 @@
+from . import auth
+from . import form
+from . import note 
+from . import profile
+
+from .auth import *
+from .form import *
+from .note import *
+from .profile import *
+
+__all__ = (
+    auth.__all__ +
+    form.__all__ +
+    note.__all__ +
+    profile.__all__
+)
