+++ api/serializers/auth.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+
+__all__ = ['TokenSerializer']
+
+
+class TokenSerializer(serializers.Serializer):
+    token = serializers.CharField()
+
